77_GPTConfigGenerator/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.0
google-generativeai>=0.3.0
pydantic>=2.5.0
pyyaml>=6.0.1
toml>=0.10.2
